app/requirements.txt
django>=4.0
psycopg2-binary>=2.9
django-redis>=5.2
djangorestframework>=3.14.0
